utils for hex to json accept file as param for ABI
