+++ CMakeLists.txt
@@ -74,12 +74,6 @@ add_executable(test_abieos_reflect src/reflect_test.cpp)
 target_include_directories(test_abieos_reflect PRIVATE include)
 add_test(NAME test_abieos_reflect COMMAND test_abieos_reflect)
 
-add_executable(generate_hex_from_json src/util_generate_hex_from_json.cpp src/abieos.cpp src/ship.abi.cpp)
-target_link_libraries(generate_hex_from_json abieos ${CMAKE_THREAD_LIBS_INIT})
-
-add_executable(generate_json_from_hex src/util_generate_json_from_hex.cpp src/abieos.cpp src/ship.abi.cpp)
-target_link_libraries(generate_json_from_hex abieos ${CMAKE_THREAD_LIBS_INIT})
-
 # Causes build issues on some platforms
 # add_executable(test_abieos_sanitize src/test.cpp src/abieos.cpp src/abi.cpp src/crypto.cpp include/eosio/fpconv.c)
 # target_include_directories(test_abieos_sanitize PRIVATE include external/outcome/single-header external/rapidjson/include external/date/include)
